requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
onnxruntime==1.20.1
pillow==11.1.0
python-multipart==0.0.20
ultralytics==8.3.75
omegaconf==2.2.2